+++ src/sc_kpm/sc_sets/__init__.py
@@ -1,0 +1,4 @@
+from sc_kpm.sc_sets.sc_numbered_set import ScNumberedSet
+from sc_kpm.sc_sets.sc_oriented_set import ScOrientedSet
+from sc_kpm.sc_sets.sc_set import ScSet
+from sc_kpm.sc_sets.sc_structure import ScStructure