+++ docs/changelog.md
@@ -12,10 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - ScAgent class for handling a single ScEvent
 - ScModule class for handling a multiple ScAgent
 - ScServer class for modules serving
+- ScSet, ScStructure, ScOrientedSet and ScNumberedSet classes for creating complex structures and manipulations
 - List of common utils for basic manipulations with knowledge base
 - List of action utils for agent development based on actions
-- List of generation utils for creating complex structures
-- List of search utils for complex construction manipulations
 - List of iterate utils for easy passing the same type of data
 - Code linting tools: isort, pylint, black, pre-commit
 - Documentation for contributors and developers